add commonMountOptions
Rather than introducing an option like `hideMounts` from impermanence,
I opted to allow for the declaration of arbitrary mount options both on
the top level of a preservation prefix and on the `users` level.

Compared to a generic `hideMounts` option, this approach would for
example allow users to choose between `x-gvfs-hide` and `x-gdu.hide`.

Closes #8